CIV-121: Default judgment form
