+++ src/main/java/uk/gov/hmcts/reform/civil/callback/CaseEvent.java
@@ -115,7 +115,8 @@ public enum CaseEvent {
     RETRY_NOTIFY_RPA_ON_CASE_HANDED_OFFLINE(CAMUNDA),
     RESET_RPA_NOTIFICATION_BUSINESS_PROCESS(CAMUNDA),
 
-    UPDATE_CASE_DATA(TESTING_SUPPORT);
+    UPDATE_CASE_DATA(TESTING_SUPPORT),
+    GENERATE_DJ_FORM(CAMUNDA);
 
     private final UserType userType;
