+++ src/main/java/uk/gov/hmcts/reform/civil/service/docmosis/DocmosisTemplates.java
@@ -11,7 +11,7 @@ public enum DocmosisTemplates {
     N9_MULTIPARTY_SAME_SOL("CV-UNS-ACK-ENG-00789.docx", "acknowledgement_of_claim_form_%s.pdf"),
     N11("CV-UNS-ACK-ENG-00799.docx", " acknowledgement_of_claim_form_%s.pdf"),
     N181_2V1("CV-UNS-HRN-ENG-00804.docx", "%s_directions_questionnaire_form_%s.pdf"),
-    N121("DJTest.docx", "default_judgment_form_%s.pdf"),
+    N121("CV-UNS-DEC-ENG-00829.docx", "default_judgment_form_%s.pdf"),
     DEFENDANT_RESPONSE_SPEC("CV-SPEC-RES-ENG-00001.docx", "%s_directions_questionnaire_form_%s.pdf");
